Contracts - completed separation of contract deployment into Contracts lib
