+++ src/contracts/Nethereum.Commerce.ContractDeployments.IntegrationTests/OverallDeploymentTests.cs
@@ -26,32 +26,32 @@ namespace Nethereum.Commerce.ContractDeployments.IntegrationTests
         {
             // If all contracts deployed and configured ok, then...
             // ...the PO storage contract should be configured to point to the eternal storage contract.
-            var actualEternalStorageAddressHeldAgainstPoStorage = await _contracts.PoStorageService.EternalStorageQueryAsync();
-            var expectedEternalStorageAddress = _contracts.EternalStorageService.ContractHandler.ContractAddress;
+            var actualEternalStorageAddressHeldAgainstPoStorage = await _contracts.Deployment.PoStorageService.EternalStorageQueryAsync();
+            var expectedEternalStorageAddress = _contracts.Deployment.EternalStorageService.ContractHandler.ContractAddress;
             actualEternalStorageAddressHeldAgainstPoStorage.Should().Be(expectedEternalStorageAddress);
 
             // ...the funding contract should be configured to point to the business partner storage contract.
-            var actualBusinessPartnerStorageAddressHeldAgainstFunding = await _contracts.FundingService.BusinessPartnerStorageQueryAsync();
-            var expectedBusinessPartnerAddress = _contracts.BusinessPartnerStorageService.ContractHandler.ContractAddress;
+            var actualBusinessPartnerStorageAddressHeldAgainstFunding = await _contracts.Deployment.FundingService.BusinessPartnerStorageQueryAsync();
+            var expectedBusinessPartnerAddress = _contracts.Deployment.BusinessPartnerStorageService.ContractHandler.ContractAddress;
             actualBusinessPartnerStorageAddressHeldAgainstFunding.Should().Be(expectedBusinessPartnerAddress);
 
             // ... the buyer wallet should be configured to point to the purchasing contract.
-            var actualPurchasingAddressHeldAgainstBuyerWallet = await _contracts.WalletBuyerService.PurchasingQueryAsync();
-            var expectedPurchasingAddress = _contracts.PurchasingService.ContractHandler.ContractAddress;
+            var actualPurchasingAddressHeldAgainstBuyerWallet = await _contracts.Deployment.WalletBuyerService.PurchasingQueryAsync();
+            var expectedPurchasingAddress = _contracts.Deployment.PurchasingService.ContractHandler.ContractAddress;
             actualPurchasingAddressHeldAgainstBuyerWallet.Should().Be(expectedPurchasingAddress);
 
             // ... the seller wallet should be configured to point to the purchasing contract.
-            var actualPurchasingAddressHeldAgainstSellerWallet = await _contracts.WalletSellerService.PurchasingQueryAsync();
+            var actualPurchasingAddressHeldAgainstSellerWallet = await _contracts.Deployment.WalletSellerService.PurchasingQueryAsync();
             actualPurchasingAddressHeldAgainstSellerWallet.Should().Be(expectedPurchasingAddress);
 
             // ... the seller wallet should be configured to have a seller id.
-            var actualSellerIdString = (await _contracts.WalletSellerService.SellerIdQueryAsync()).ConvertToString();
+            var actualSellerIdString = (await _contracts.Deployment.WalletSellerService.SellerIdQueryAsync()).ConvertToString();
             var expectedSellerIdString = _contracts.Deployment.ContractDeploymentConfig.EShopSellerId;
             actualSellerIdString.Should().Be(expectedSellerIdString);
 
             // ... and that seller id should have a master data entry in business partner storage.            
             var actualSellerIdBytes = actualSellerIdString.ConvertToBytes();
-            var actualSellerIdRecordFromBusinessPartnerStorage = (await _contracts.BusinessPartnerStorageService.GetSellerQueryAsync(actualSellerIdBytes)).Seller;
+            var actualSellerIdRecordFromBusinessPartnerStorage = (await _contracts.Deployment.BusinessPartnerStorageService.GetSellerQueryAsync(actualSellerIdBytes)).Seller;
             actualSellerIdRecordFromBusinessPartnerStorage.IsActive.Should().Be(true);
             actualSellerIdRecordFromBusinessPartnerStorage.SellerDescription.Should().Be(_contracts.Deployment.ContractDeploymentConfig.EShopDescription);
         }
@@ -60,9 +60,9 @@ namespace Nethereum.Commerce.ContractDeployments.IntegrationTests
         public async void ShouldHaveDeployedMockContracts()
         {
             // If mock contracts deployed ok, then a MockDai contract should exist
-            var totalSupply = await _contracts.MockDaiService.TotalSupplyQueryAsync();
+            var totalSupply = await _contracts.Deployment.MockDaiService.TotalSupplyQueryAsync();
             totalSupply.Should().BeGreaterThan(1);
-            var dec = await _contracts.MockDaiService.DecimalsQueryAsync();
+            var dec = await _contracts.Deployment.MockDaiService.DecimalsQueryAsync();
             dec.Should().BeGreaterThan(0);
             var totalSupplyFactored = totalSupply / BigInteger.Pow(10, dec);
             _output.WriteLine($"MockDai Total Supply = {totalSupplyFactored.ToString("N0")}");
