+++ src/contracts/Nethereum.Commerce.ContractDeployments.IntegrationTests/appsettings.json
@@ -1,5 +1,5 @@
 {
-  "NewDeployments": {
+  "NewDeployment": {
     "BlockchainUrl": "http://testchain.nethereum.com:8545",
     "EShopSellerId": "Nethereum.eShop",
     "EShopDescription": "Satoshi's Books",
